zstd: histogram

+++ src/main/java/io/airlift/compress/zstd/huffman/HuffmanCompressionTable.java
@@ -444,6 +444,6 @@ public final class HuffmanCompressionTable {
         }
         output += compressedSize;
 
-        return (int) (output - outOffs);
+        return output - outOffs;
     }
 }
